requirements.txt
boto3
pandas
apache-airflow
pytest
